Search Insert Pos:
Given a sorted array and a target value, return the index if the
target is found. If not, return the index where it would be
if it were inserted ir order.

My solution is not the most efficient at O(N).

Discussion solution uses binary search which is optimal solution
O(lgn)